requirements.txt
selenium>=4.8.0
typer>=0.7.0
mergedeep>=1.3.4
pyyaml>=23.0.1